ci: use maintained action to setup rust toolchain (#585)
## What changes are proposed in this pull request?

The `actions-rs/toolchain` action is deprecated in favor of
`actions-rust-lang/setup-rust-toolchain`. This PR updates the usages of
the respective actions in the github workflows. THe new action already
includes an integration with the rust-cache action, so no need to set
that up separately anymore.

This also sets up a dependabot configuration for `cargo` and
`github-actions` which we may or may not choose to keep.

## How was this change tested?

no code changes.

---------

Co-authored-by: Zach Schuermann <zachary.zvs@gmail.com>