refactor: extract deduplication logic from `AddRemoveDedupVisitor` into embeddable `FileActionsDeduplicator` (#769)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

<!--
PR title formatting:
This project uses conventional commits:
https://www.conventionalcommits.org/

Each PR corresponds to a commit on the `main` branch, with the title of
the PR (typically) being
used for the commit message on main. In order to ensure proper
formatting in the CHANGELOG please
ensure your PR title adheres to the conventional commit specification.

Examples:
- new feature PR: "feat: new API for snapshot.update()"
- bugfix PR: "fix: correctly apply DV in read-table example"
-->

## What changes are proposed in this pull request?
<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->

**No behavioral changes were introduced, this is purely a refactoring
effort**

This PR extracts the core deduplication logic from the
`AddRemoveDedupVisitor` in order to be shared with the incoming
`CheckpointVisitor`. For a bigger picture view on how this refactor is
helpful, please take a look at the following PR which implements the
`CheckpointVisitor` with an embedded `FileActionsDeduplicator` that will
rebase this PR once merged. [[link to
PR]](https://github.com/delta-io/delta-kernel-rs/pull/738).

This `FileActionsDeduplicator` lives in the new top-level `log_replay`
mod as it will be leveraged in the nested `scan/log_replay` mod and the
incoming `checkpoints/log_replay` mod. There are also additional traits
& structs that the two `log_replay` implementations will share via this
new top-level mod. For an even wider view of the implementation of the
`checkpoints` mod and the component re-use, please have a look at the
following PR. [[link to
PR]](https://github.com/delta-io/delta-kernel-rs/pull/744)

## Summary of refactor
1. New `log_replay` mod
2. Moved `FileActionKey` definition from `scan/log_replay` to the new
`log_replay` mod
3. New `FileActionDeduplicator` in the new `log_replay` mod
- Includes the `check_and_record_seen` method which was simply **moved**
from the `AddRemoveDedupVisitor`
- Includes the `extract_file_action` method and `extract_dv_unique_id`
private method which may be new concepts, but includes functionality
which are both pieces of functionality pulled from the
`AddRemoveDedupVisitor` to be shared with the incoming
`CheckpointVisitor`



<!--
Uncomment this section if there are any changes affecting public APIs:
### This PR affects the following public APIs

If there are breaking changes, please ensure the `breaking-changes`
label gets added by CI, and describe why the changes are needed.

Note that _new_ public APIs are not considered breaking.
-->


## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->

All existing tests pass ✅