refactor: extract log replay processing structure into `LogReplayProcessor` trait (#774)
<!--
Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines:
https://github.com/delta-incubator/delta-kernel-rs/blob/main/CONTRIBUTING.md
2. Run `cargo t --all-features --all-targets` to get started testing,
and run `cargo fmt`.
  3. Ensure you have added or run the appropriate tests for your PR.
4. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP]
Your PR title ...'.
  5. Be sure to keep the PR description updated to reflect all changes.
-->

<!--
PR title formatting:
This project uses conventional commits:
https://www.conventionalcommits.org/

Each PR corresponds to a commit on the `main` branch, with the title of
the PR (typically) being
used for the commit message on main. In order to ensure proper
formatting in the CHANGELOG please
ensure your PR title adheres to the conventional commit specification.

Examples:
- new feature PR: "feat: new API for snapshot.update()"
- bugfix PR: "fix: correctly apply DV in read-table example"
-->

**No behavioral changes were introduced, this is purely a refactoring
effort**

## What changes are proposed in this pull request?
<!--
Please clarify what changes you are proposing and why the changes are
needed.
The purpose of this section is to outline the changes, why they are
needed, and how this PR fixes the issue.
If the reason for the change is already explained clearly in an issue,
then it does not need to be restated here.
1. If you propose a new API or feature, clarify the use case for a new
API or feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->

This PR refactors the newly-named `ScanLogReplayProcessor` to implement
the newly introduced `LogReplayProcessor` trait in order to unify
components for the incoming `CheckpointLogReplayProcessor`. For a bigger
picture view of how this refactor is helpful, please have a look at the
following PR, which introduces the `CheckpointLogReplayProcessor` which
implements the `LogReplayProcessor` trait, that will rebase this PR once
merged #744


### Summary of changes
- Introduced the `LogReplayProcessor` trait in the top-level
`log_replay` mod
- Renamed `LogReplayScanner` in scan/log_replay ->
`ScanLogReplayProcessor`
- Updated `ScanLogReplayProcessor` to implement the `LogReplayProcessor`
trait
- requires pushing down the `add_transform`, `logical_schema`, and
`transform` fields to the `ScanLogReplayProcessor` for accessibility in
the `process_actions_batch` trait method, as the trait method
implementation has fixed parameters.
- Simplified the `scan_action_iter` function to use the trait's
`apply_to_iterator` method

<!--
Uncomment this section if there are any changes affecting public APIs:
### This PR affects the following public APIs

If there are breaking changes, please ensure the `breaking-changes`
label gets added by CI, and describe why the changes are needed.

Note that _new_ public APIs are not considered breaking.
-->


## How was this change tested?
<!--
Please make sure to add test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please
clarify how you tested, ideally via a reproducible test documented in
the PR description.
-->

All existing tests pass ✅