chore!: update dependencies (#814)
title, minimal changes to integrate some breaking changes. note that
this is not a breaking change for kernel since the types from crates
with breaking changes never escape the kernel.

breaking updates (remaining updates are all non-breaking):
- itertools 0.13 -> 0.14
- itertools 0.14 increases MSRV and removes a generic param from a
function we don't use
- thiserror 1 -> 2
- [thiserror 2](https://github.com/dtolnay/thiserror/releases/tag/2.0.0)
seems to impose restrictions which don't affect us.
- strum 0.26 -> 0.27
- two breaking update for their release (1) MSRV increase to 1.66 and
(2) something to do with `EnumProperties` which we don't use.
- (build dep) ureq 2.10 -> 3
- (dev dep) datatest-stable 0.2 -> 0.3