refactor: new `#[internal_api]` macro instead of `visibility` crate (#835)
## What changes are proposed in this pull request?
In an effort to simplify our system of having an `internal_api` API this
PR introduces an `#[internal_api]` attribute to annotate functions that
are `pub(crate)` to make them `pub` when the `internal_api` feature flag
is enabled. In this process, the PR removes our dependency on
`visibility` crate (and hopefully makes it easier to introduce/track our
'internal' APIs). Note that there's also a small `internal_mod!` macro
which is used to workaround the limitation of rustc disallowing
attributes on non-inline modules.

~Stacked on #834~

## How was this change tested?
build