fix: disable predicate pushdown in `Scan::execute` (#861)
## What changes are proposed in this pull request?
TLDR: Disable predicate pushdown in `Scan::execute` API (and CDF API
too) since it currently breaks DV application.

The `Scan::execute` function currently combines the
`Scan::scan_metadata()` (and the needed visiting, etc.) to get the
`ScanFiles`, then for each scan file, performs the corresponding DV and
parquet read before applying transforms/schemas and eventually handing
back 'final data' to the user (in the form of a `ScanResult` which is
just filtered engine data).

The problem lies in the application of the DV to the data in the
presence of predicates. currently everything is positional, that is, if
we read a DV that indicates rows 1, 5, and 10 are deleted from a record
batch, then `Scan::execute` will simply apply those positional deletions
after reading the parquet file. This may seem okay at first glance but
if a predicate 'reshuffles' data from the parquet file then the
positions referenced in the aforementioned DV are no longer valid. This
should instead be handled by row indexes. future work: #860

## How was this change tested?
Updated existing `test_data_row_group_skipping` UT to assert there is
_no_ predicate pushdown when using `scan.execute`