fix: allow CheckpointDataIterator to be used across await (#961)
## What changes are proposed in this pull request?

Writing data is mostly an async operation in Rust. Since the
`CheckpointDataIterator` needs to be handed back to kernel on
`finalize`, we need to be able to use it across `awaits`. As such we
require the internal boxed iterator to also be `Send`.

## How was this change tested?

This came up when integrating checkpoint writing in `delta-rs`. With the
proposed change, it works.