feat(ffi): add timetravel by version number (#1044)
## What changes are proposed in this pull request?
- Add a new ffi function called `snapshot_at_version`, which allows
getting a specific snapshot of a table.

While this is by itself not super useful to users yet (there's not
really a way to list snapshots through the ffi), it will allow testing
timetravel in the DuckDB delta extension. A follow up to this PR could
be to add:
- a function to list snapshots
- a function that gets a snapshot that corresponds to a timestamp

### This PR affects the following public APIs
- `snapshot_at_version` added to the ffi

## How was this change tested?
using a few extra lines that are added to the existing snapshot ffi
function tests.

---------

Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>
Co-authored-by: Zach Schuermann <zachary.zvs@gmail.com>