feat: Support reordering structs that are inside maps in default parquet reader (#1060)
## What changes are proposed in this pull request?
Previously we didn't reorder things that were inside a map. We actually
already did compute the required reordering, but then just threw it
away.

This uses that computed reordering and returns it in the mask/reorder
phase, and then uses it in `reorder_struct_array` to properly recurse
into maps and reorder elements.

We model a map reorder as a `ReorderIndex::Nested` variant. The index of
the top level nested item says where the map itself should end up in the
output. Then there are two nested `ReorderIndex`s, which specify how any
inner things should be reordered. The code does handle the case that a
struct might be in a map but not need to be reordered. It checks if the
computed nested reorder for keys or values doesn't actually require any
work and will replace it with an `Identity` transform if none is needed.

## How was this change tested?
Unit tests. One for computing the reorder, one for applying it.