feat: Add default engine support for arrow eval of opaque expressions (#980)
## What changes are proposed in this pull request?

Follow-up to 
* https://github.com/delta-io/delta-kernel-rs/pull/686

Add support for arrow evaluation of opaque expressions in the default
engine. The new API allows engines to define their own expressions that
operate on arrow record batches and which seamlessly fit into the arrow
evaluation framework alongside normal kernel expressions.

Once this support lands, we can remove the existing
`BinaryPredicateOp::In` predicate from the core kernel, and make it part
of the default engine instead.

### This PR affects the following public APIs

Add a new API for evaluating opaque expressions and predicates in the
default engine.

## How was this change tested?

New unit tests.