feat(ffi): added default-engine-rustls feature and extern "C" for .h file (#1023)
This PR adds a feature `default-engine-rustls` to allow users to use the
feature in the kernel. Additionally, it adds `cpp_compat = true` to the
`cbindgen.toml` file so that it will add a section to add the `extern
"C"` wrapper.
```c
#ifdef __cplusplus
extern "C" {
#endif // __cplusplus
```


## What changes are proposed in this pull request?

1. This commit updates the generated `delta_kernel_ffi.h` to include c++
compatibility to that it will include `extern "C"` so that code
generation will appropriately not try to use mangled method names.
2. Added the feature `delta-engine-rustls` to the ffi project in order
to use the same feature from the kernel.


## How was this change tested?

Tested against the dotnet implementation of delta lake.

---------

Co-authored-by: Nick Lanham <nicklan@users.noreply.github.com>