test: add unshredded variant read test (#1088)
## What changes are proposed in this pull request?
Add a test where we read a table with lots of unshredded variants

Note: this table has quite a few rows. I've added the expected output as
its own file. We could regenerate a smaller table, but it seemed okay to
use this as it's already been generated by the variant team.

## How was this change tested?
It is a test :)