fix: VARIANT follow-ups for SchemaTransform etc (#1106)
## What changes are proposed in this pull request?
Follow-up from VARIANT work in #1015. chiefly: (1) modify
`transform_variant` to take/return structtypes, and default to recursing
through children, and (2) track a couple follow ups in TODOs.

the actual bug that must be fixed is that when a VARIANT went in to
`transform_variant`, it came back out as a regular STRUCT if
transformed, since DataType::from on a struct is just a struct (note
that if not transformed, we return the original data_type which is still
variant)

original comment kept below for posterity

### Original exploration in earlier commits
Originally prompted as follow-up from #1015, this serves as an
exploration of introducing a new `VariantType` struct (similar to
`MapType`/`ArrayType`/etc.) which just wraps `StructType`.

The original question pertaining to `SchemaTransform` was basically: can
we implement `transform_variant` to take+return a `StructType` instead
of `DataType`? This loses the information that says the thing is a
variant and not a normal struct. This prompted the exploration of how we
actually want to represent Variant in our schemas. In general we need a
way of distinguishing variants from plain structs

I think we have two options: we either rely on `DataType` which
communicates `DataType::Variant(struct_type)` OR we introduce a new
`StructType` much like MapType/ArrayType/etc.

This PR serves as exploration of the former - though i'm realizing if we
_do_ take `DataType` as the thing that 'tags' a struct as variant then
it's probably also okay? And would just have the schema transform still
retain taking/returning DataType? (or perhaps could consume a StructType
but just always return a DataType)

I remember @nicklan had even more/better thoughts around this - please
comment and we can discuss :)

## How was this change tested?
Existing UT