feat: add append functions to ffi (#962)
This PR is a retake of
https://github.com/delta-io/delta-kernel-rs/pull/550

## What changes are proposed in this pull request?
This PR implements the ffi functions required to do appends. The main
difference over https://github.com/delta-io/delta-kernel-rs/pull/550 is
how it's tested: instead of using a C based example we now opt for a
rust-based test that is able to re-used various testing-related utility
functions that are used in the kernel tests. To do this cleanly, i've
moved several functions over to the test-utils crate. 
## This PR affects the following public APIs
- Adds new ffi functions related to appends / transactions

## How was this change tested?
Using the `test_basic_append` in `ffi/src/transaction/mod.rs`. 


## Main points to comment on for reviewers of Draft PR:
I would like to draw the reviewer's attention to:
- the concept of testing new ffi functionality through rust tests
instead of a C example program
- usage of `tempfile::tempdir` to create a tmp directory for test data
during the test
- some outstanding `TODO`s in the `test_basic_append` test

Once the current approach is approved by reviewers, I can follow up with
more testing:
- the write schema
- appending to a partitioned table