feat!: `make_physical` takes column mapping and sets parquet field ids (#1082)
## What changes are proposed in this pull request?

When the column mapping mode is set to Id, the `make_physical` function
will now insert the parquet field id into the schema. This will be used
in the future to facilitate column mapping mode for parquet reads.

To show that ParqueFieldId is the correct metadata to read to get the
field Id see the following:
-
[delta-spark](https://github.com/delta-io/delta/blob/master/spark/src/main/scala/org/apache/spark/sql/delta/DeltaColumnMapping.scala#L245)
-
[java-kernel](https://github.com/delta-io/delta/blob/master/kernel/kernel-defaults/src/main/java/io/delta/kernel/defaults/internal/parquet/ParquetSchemaUtils.java#L404)

# Is this a breaking change?

Yes. `StructField::make_physical` is now an internal_api instead of a
public function. Its signature has also changed.

In ColumnMappingMode is None, then the physical schema's name is the
logical name. Previously, we would unconditionally use the column
mapping physical name, even if column mapping mode is none.



## How was this change tested?
Ensure that parquet field ids are included in the physical schema only
when column mapping mode is set to Id, and that it works for nested
types.