refactor: new listed_log_files module (#1150)
## What changes are proposed in this pull request?
`ListedLogFiles` was introduced as a means to abstract the files listed
from the log. Though we still retained many static methods for listing
and all the code lived in `mod log_segment`. This PR refactors
`ListedLogFiles` to do the following (each number in this list
corresponds to a commit in the chain - per-commit review encouraged)
1. Move `ListedLogFiles` code from `mod log_segment` to a new `mod
listed_log_files`
2. Replace static method `list_log_files(...)` with
`ListedLogFiles::list_commits(...)`
3. Replace static methods `list_log_files_with_version(...)` with
`ListedLogFiles::list(...)` and `list_log_files_with_checkpoint(...)`
with `ListedLogFiles::list_with_checkpoint_hint(...)`
4. Remove `ListedLogFiles::try_new(...)` from `pub(crate)` API ->
everything must go through the new constructors above. (note we still
allow it for `cfg(test)`)

Note that there is a relatively large diff - this is mostly due to code
movement. First commit is entirely code movement (and it labeled as
such).

## How was this change tested?
modified existing UT

---------

Co-authored-by: OussamaSaoudi <45303303+OussamaSaoudi@users.noreply.github.com>